feat: extract Mayo heckler engine into shared module
mayo-heckler.mjs is now the single source of truth for all heckle
data, escalation logic, agent comebacks, and explosion sequences.
agent-logs.mjs imports and wraps with ANSI formatting.
Agent Monitor SPA has its own inline copy (browser can't import .mjs
from Node paths) but uses the same event schema.
